+++ h2integrate/control/test/test_openloop_controllers.py
@@ -1,21 +1,20 @@
 from copy import deepcopy
 from pathlib import Path
 
-import yaml
 import numpy as np
 import pytest
 import openmdao.api as om
 from pytest import fixture
-from openmdao.utils.assert_utils import assert_check_totals
 
-from h2integrate.control.control_strategies.passthrough_openloop_controller import (
-    PassThroughOpenLoopController,
-)
+from h2integrate.core.file_utils import load_yaml
 from h2integrate.control.control_strategies.storage.demand_openloop_controller import (
     DemandOpenLoopStorageController,
 )
 from h2integrate.control.control_strategies.converters.demand_openloop_controller import (
     DemandOpenLoopConverterController,
+)
+from h2integrate.control.control_strategies.storage.passthrough_openloop_controller import (
+    PassThroughOpenLoopController,
 )
 from h2integrate.control.control_strategies.converters.flexible_demand_openloop_controller import (
     FlexibleDemandOpenLoopConverterController,
@@ -47,11 +46,15 @@ def test_pass_through_controller(subtests):
     tech_config_path = current_dir / "inputs" / "tech_config.yaml"
 
     # Load the technology configuration
-    with tech_config_path.open() as file:
-        tech_config = yaml.safe_load(file)
+    tech_config = load_yaml(tech_config_path)
 
+    tech_config["technologies"]["h2_storage"]["model_inputs"]["shared_parameters"].update(
+        {"set_demand_as_avg_commodity_in": True}
+    )
     # Set up the OpenMDAO problem
     prob = om.Problem()
+
+    plant_config = {"plant": {"plant_life": 30, "simulation": {"n_timesteps": 10}}}
 
     prob.model.add_subsystem(
         name="IVC",
@@ -62,7 +65,7 @@ def test_pass_through_controller(subtests):
     prob.model.add_subsystem(
         "pass_through_controller",
         PassThroughOpenLoopController(
-            plant_config={}, tech_config=tech_config["technologies"]["h2_storage"]
+            plant_config=plant_config, tech_config=tech_config["technologies"]["h2_storage"]
         ),
         promotes=["*"],
     )
@@ -73,26 +76,9 @@ def test_pass_through_controller(subtests):
 
     # Run the test
     with subtests.test("Check output"):
-        assert np.arange(10) == pytest.approx(
-            prob.get_val("hydrogen_set_point", units="kg/h"), rel=1e-3
-        )
-
-    # Run the test
-    with subtests.test("Check derivatives"):
-        # check total derivatives using OpenMDAO's check_totals and assert tools
-        assert_check_totals(
-            prob.check_totals(
-                of=[
-                    "hydrogen_set_point",
-                ],
-                wrt=[
-                    "hydrogen_in",
-                ],
-                step=1e-6,
-                form="central",
-                show_only_incorrect=False,
-                out_stream=None,
-            )
+        expected_set_point = np.mean(np.arange(10)) - np.arange(10)
+        assert expected_set_point == (
+            pytest.approx(prob.get_val("hydrogen_set_point", units="kg/h"), rel=1e-3)
         )
 
 
@@ -105,8 +91,7 @@ def test_storage_demand_controller(subtests):
     tech_config_path = current_dir / "inputs" / "tech_config.yaml"
 
     # Load the technology configuration
-    with tech_config_path.open() as file:
-        tech_config = yaml.safe_load(file)
+    tech_config = load_yaml(tech_config_path)
 
     plant_config = {"plant": {"simulation": {"n_timesteps": 10}}}
 
@@ -182,8 +167,7 @@ def test_storage_demand_controller_round_trip_efficiency(subtests):
     tech_config_path = current_dir / "inputs" / "tech_config.yaml"
 
     # Load the technology configuration
-    with tech_config_path.open() as file:
-        tech_config = yaml.safe_load(file)
+    tech_config = load_yaml(tech_config_path)
 
     plant_config = {"plant": {"simulation": {"n_timesteps": 10}}}
 
@@ -276,8 +260,7 @@ def test_storage_demand_controller_round_trip_with_non_one_efficiencies(subtests
     tech_config_path = current_dir / "inputs" / "tech_config.yaml"
 
     # Load the technology configuration
-    with tech_config_path.open() as file:
-        tech_config = yaml.safe_load(file)
+    tech_config = load_yaml(tech_config_path)
 
     plant_config = {"plant": {"simulation": {"n_timesteps": 10}}}
 
@@ -421,8 +404,7 @@ def test_generic_storage_demand_controller(subtests):
     tech_config_path = current_dir / "inputs" / "tech_config.yaml"
 
     # Load the technology configuration
-    with tech_config_path.open() as file:
-        tech_config = yaml.safe_load(file)
+    tech_config = load_yaml(tech_config_path)
 
     tech_config["technologies"]["h2_storage"] = {
         "performance_model": {
@@ -508,8 +490,7 @@ def test_demand_converter_controller(subtests):
     tech_config_path = current_dir / "inputs" / "tech_config.yaml"
 
     # Load the technology configuration
-    with tech_config_path.open() as file:
-        tech_config = yaml.safe_load(file)
+    tech_config = load_yaml(tech_config_path)
 
     tech_config["technologies"]["load"] = {
         "control_strategy": {
@@ -576,8 +557,7 @@ def test_flexible_demand_converter_controller(subtests, variable_h2_production_p
     tech_config_path = current_dir / "inputs" / "tech_config.yaml"
 
     # Load the technology configuration
-    with tech_config_path.open() as file:
-        tech_config = yaml.safe_load(file)
+    tech_config = load_yaml(tech_config_path)
 
     end_use_rated_demand = 10.0  # kg/h
     ramp_up_rate_kg = 4.0
@@ -687,8 +667,7 @@ def test_flexible_demand_converter_controller_min_utilization(
     tech_config_path = current_dir / "inputs" / "tech_config.yaml"
 
     # Load the technology configuration
-    with tech_config_path.open() as file:
-        tech_config = yaml.safe_load(file)
+    tech_config = load_yaml(tech_config_path)
 
     end_use_rated_demand = 10.0  # kg/h
     ramp_up_rate_kg = 4.0